pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "prism"
version = "0.3.0"
description = "Modular, deterministic, colour-aware data-quality orchestrator."
readme = "README.md"
requires-python = ">=3.9"
license = { file = "LICENSE" }
authors = [
    { name = "Daniel Austin", email = "the.potato.gnome@gmail.com" }
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent"
]
dependencies = [
    "tUilKit>=0.5.1",
    "pyyaml>=6.0"
]

[project.scripts]
prism = "prism.interfaces.cli.cli:main"

[project.urls]
Repository = "https://github.com/potatognome/dev_local"

[tool.setuptools.packages.find]
where = ["src"]
include = ["prism*"]


